fix(@eurora/web): vcard routes not compiling
