+++ apps/web/svelte.config.js
@@ -15,7 +15,7 @@ const config = {
 			assets: 'dist',
 			fallback: null,
 			precompress: true,
-			strict: true
+			strict: true,
 		})
 	},
 	compilerOptions: {
